fix: 都道府県定数を設定に依存しないモジュールへ切り出す
#73 で build-merged-csv.js に PREFECTURE_NAMES を lib/normalize.js から
import したところ、クローラーの配信物バリデーションが起動時に落ちた。

    Error: sources.yaml が見つかりません: /app/config/sources.yaml
        at loadConfig (config.js:38)
        at lib/normalize.js:12

lib/normalize.js はトップレベルで loadConfig() を実行する。クローラーは
sources.yaml をコミットせず、クロール実行時だけ CONFIG_PATH で公開リポの
クローンを指す設計のため、その環境変数を渡さない `node scripts/test.js`
（配信物バリデーション）から到達すると設定ファイルが見つからず停止する。

設定に依存しない都道府県の定数を lib/prefectures.js に切り出し、
normalize.js からは再エクスポートして既存の import を保つ。
build-merged-csv.js は新しいモジュールを参照する。

あわせて、scripts/test.js と build-merged-csv.js から lib/config.js へ
到達できないことを、import を静的にたどるユニットテストで固定する。
