+++ scripts/lib/normalize.js
@@ -11,23 +11,12 @@ import { loadConfig } from './config.js';
 // mapRecord / detectHeaderRow はこれを既定で使う（テストは引数なしで呼ぶ）。
 const DEFAULT_COLUMN_MAP = loadConfig().columnMap;
 
-// ---------------------------------------------------------------------------
-// JIS都道府県コード（先頭2桁）→ 都道府県名
-// ---------------------------------------------------------------------------
-export const PREFECTURE_BY_CODE = {
-  '01': '北海道', '02': '青森県', '03': '岩手県', '04': '宮城県', '05': '秋田県',
-  '06': '山形県', '07': '福島県', '08': '茨城県', '09': '栃木県', '10': '群馬県',
-  '11': '埼玉県', '12': '千葉県', '13': '東京都', '14': '神奈川県', '15': '新潟県',
-  '16': '富山県', '17': '石川県', '18': '福井県', '19': '山梨県', '20': '長野県',
-  '21': '岐阜県', '22': '静岡県', '23': '愛知県', '24': '三重県', '25': '滋賀県',
-  '26': '京都府', '27': '大阪府', '28': '兵庫県', '29': '奈良県', '30': '和歌山県',
-  '31': '鳥取県', '32': '島根県', '33': '岡山県', '34': '広島県', '35': '山口県',
-  '36': '徳島県', '37': '香川県', '38': '愛媛県', '39': '高知県', '40': '福岡県',
-  '41': '佐賀県', '42': '長崎県', '43': '熊本県', '44': '大分県', '45': '宮崎県',
-  '46': '鹿児島県', '47': '沖縄県',
-};
+// 都道府県の定数は lib/prefectures.js に置き、ここからは再エクスポートする
+// （このモジュールは冒頭で loadConfig() を実行するため、設定を必要としない
+// 利用者が都道府県名だけを import できるようにするため）。
+import { PREFECTURE_BY_CODE, PREFECTURE_NAMES } from './prefectures.js';
 
-export const PREFECTURE_NAMES = Object.values(PREFECTURE_BY_CODE);
+export { PREFECTURE_BY_CODE, PREFECTURE_NAMES };
 
 // 住所文字列から都道府県・市区町村を切り出す。
 //   例: 「北海道足寄郡足寄町南三条…」→ ['北海道','足寄郡足寄町']
