+++ scripts/lib/prefectures.js
@@ -1,0 +1,27 @@
+// 都道府県の定数。
+//
+// ここに置く理由: normalize.js はトップレベルで loadConfig()（config/sources.yaml の
+// 読み込み）を実行するため、都道府県名がほしいだけのモジュールが normalize.js を
+// import すると、設定ファイルが無い環境では起動時に落ちる。クローラーの配信物
+// バリデーションは CONFIG_PATH を渡さずに走るため、実際にこれで停止した。
+// 設定に依存しない定数はこのモジュールに独立させ、どこからでも安全に import できる
+// ようにする（normalize.js からは再エクスポートするので既存の import はそのまま動く）。
+
+// ---------------------------------------------------------------------------
+// JIS都道府県コード（先頭2桁）→ 都道府県名
+// ---------------------------------------------------------------------------
+export const PREFECTURE_BY_CODE = {
+  '01': '北海道', '02': '青森県', '03': '岩手県', '04': '宮城県', '05': '秋田県',
+  '06': '山形県', '07': '福島県', '08': '茨城県', '09': '栃木県', '10': '群馬県',
+  '11': '埼玉県', '12': '千葉県', '13': '東京都', '14': '神奈川県', '15': '新潟県',
+  '16': '富山県', '17': '石川県', '18': '福井県', '19': '山梨県', '20': '長野県',
+  '21': '岐阜県', '22': '静岡県', '23': '愛知県', '24': '三重県', '25': '滋賀県',
+  '26': '京都府', '27': '大阪府', '28': '兵庫県', '29': '奈良県', '30': '和歌山県',
+  '31': '鳥取県', '32': '島根県', '33': '岡山県', '34': '広島県', '35': '山口県',
+  '36': '徳島県', '37': '香川県', '38': '愛媛県', '39': '高知県', '40': '福岡県',
+  '41': '佐賀県', '42': '長崎県', '43': '熊本県', '44': '大分県', '45': '宮崎県',
+  '46': '鹿児島県', '47': '沖縄県',
+};
+
+/** 47都道府県の名前（JISコード順）。 */
+export const PREFECTURE_NAMES = Object.values(PREFECTURE_BY_CODE);
